09/29/16 12:26AM
+ "InputManager" script which keeps track of all the various input
buttons / joysticks the programmer wishes to use and their settings.
This script will continually check if any of these buttons / joysticks
go from inactive to active and vice versa.

+ "ControllableObject" script maintains a dictionary of all the keys
which have been pressed. This allows the system to determine if a key
has been pressed / released and for the duration of that time inbetween.

+ "AbstractAction" script was designed as a base for all future actions.
It contains several GameObjects which will be necessary for every
action.

+ "FacingDirection" script extends "AbstractAction" and  simply changes
the variable in "ControllableObject" from left or right, depending on
the input.

+ "Walking" script extends "AbstractAction" and allows the player to
walk left and right on the X-Axis.