11/02/2016 3:12AM
* Updated CameraShake to it properly it always has the correct position
when moving to a new section of the level and screen shaking
* Modified AbstractGun to be a true abstraction for all weapons. This
will be better for other weapons if we decide to make them. Also, this
is how it shoud have been coded in the first place. : D
* Updated Shotgun class to properly implement the abstractgun class and
have unique behaviros
* Updated MachineGun class to properly implement the abstractgun class
and have unique behaviors
* Placed all animations in the animations folder
* Added the Physics2D.PlatformEffector2D on all the floating platforms
in the first area. This saves us a ton of work for platforms.