11/12/2016 7:15PM
+ Added an array in ControllableObject script which maintains an array
of angles and the current key for the angle currently being used. This
was implemented to greatly reduce the amount of if statements we had
scattered trhoughout various scripts.
* Added an array of functions in the base class for player weapons. This
array uses the current index from controllable object to determine which
function to call. This has an opertaion of O(1) vs. checking if
statements throughout various scripts.
* Fixed the crystal bullet rotation issue with this newly implemented
array of angles in controllable object. checks the current index and
uses the the angle stored int it to correct its rotation.