11/23/2016 7:33PM
* Updated the reload system for both the machine gun and shotguns. There
is a fast reload if the player was in the air and then lands. There is a
slower reload for when the player is on the ground the entire time, but
it is actually slightly faster than if the player was to shoot
themselves into the air just to get the fast reload. This prevents the
players from wanting to do one reload over the other.
* Fixed an issue where the player was not lifting with the machine gun
under certain situations.
* Fixed a a bug with the ammos not displaying properly.