12/18/2016 7:04PM
+ Added a Screen Shake Manager: The shotgun, machine, and player health
currently use this new system. They send a message to the manager with
the type of screen shake they want. The request is then put into a queue
and processed in the order in which they came. The manager is also in
charge of bringing the local position back to Vector3.zero
+ Added a chromatic abberation effect when the player is hit. This is
effect/shader is brought in through Unity's standard assets. This is why
there are so many changes in this commit.
* Modified the smart camera into 3 sections: FOV, X-plane, and Y-plane.
Each script is in charge of determining how the camera should behave in
a given situation.