12/20/2016 3:18AM
+ Added a Movement Manager for the player, similar to the screen shake
manager. There are scriptableObjects which are used in the
playerWalking, shotgun, and machine gun script. These scriptableObjects
are used to request movement from the manager. The manager then puts
each request in a queue. When it is time to dequeue each request, the
manager passes information such as current aim key (used for direction),
current velocity, and collision detection. The scriptableObject will
take that information and decide how it should be processed. Lastly, it
returns back a Vector2 to the manager so it may be applied to the
player.
* The new movement system has been applied all user controlled player
movement scripts.
- Deleted many of the other scripts which have become invalid