01/08/2017 1:38AM
* Added Spriter Pro Game Effects to the project. We will not be using
all of these, but I imported all of them to play around with the ones we
like.
+ Added a dust effect when the player lands.
* Fixed and finished the focus camera. This is where the player reaches
an area  we want the camer to be fixed. After the player has left the
area, the camera's normal functions resume.
* All animations for the player can now be found in one scriptable
object. To use an animation, the player just needs to create a new
instance of this SO, attach it to the desired behavior and call
"PlayAnimation()" when they want.
* Many, many, many other changes were made, but they were all to make
things easier to find or change for other users. No functionality was
changed for these changes.