01/14/2017 4:42PM
* Updated how player shots interact with and damage enemy units. The
bullets will now directly call the DecrementHealth function for only the
enemy in which it hit. Much more efficient.
+ Created SOEffects script, which is used to create an Asset called
SOEffectsHandler. Scripts can use this single SO and tell it which
effect to play with optional parameters.
This one SO will have all the available effects in the game in one
location and only take up space one time.
* Deleted all the assets from Spriter Pro which we were not using.
Importing on a needed basis from now on.