01/23/2017 12:54AM
+ Completed the Acid Swarmer enemy. This includes: shooting a green acid
ball which damages the player, the splatter from the acid ball when it
collides with surfaces, and the shoot out effect when the ball is
instantiate.
+ Create a acid damaging effect when the player is hit by the acid ball.
This includes an effect animation which plays above the player's head
and turns the sprite's color green. Over a specified duration, the color
returns to white.
+ Completed the Explosive Swarmer enemy. This includes the effects
animation to help differentiate it, the player function call, and a few
other details.
+ Player has an explosive damaging effect which gives the player a
char'd look and returns to white over a specified time period.
* Other various additions, tweaks, and changes.