02/13/2017 12:35PM
* Fixed a couple of bugs with the shotgun blast. It no longer goes
through certain object. It always stops upon collision and plays the
lighting impact effect.
* Fixed the shotgun blast with the charger enemy. The blast no longer
damages the charger enemy from the front.
* Fixed the crystal bullet with the charger enemy. The bullet no longer
damages the charger enemy from the front.
* Fixed and issue where both the shotgun blast and the crystal bullet
went through the bottom of platforms. Added a boxcollider2d as a trigger
to notify the ammuntion not to go through.
* Fixed an issue where the charger enemy would charge even if the player
was not on the same y-axis. It now only charges when the play is in
"sights".
* A few other minor changes here and there to improve the charger enemy.