02/19/2017 1:32AM
* Updated the player health script so the respawn animation plays.
* Updated the player respawn behaviors. Now checkpoints will be needed
throughout the game. These will be the only spots where players can be
respawned.
* Created a script which enables and disables the smart camera features.
This is to so we, as developers, can turn off the smart camera features
and move the camera exactly the way we envision. Have a dynamic camera
proved to be more of challenge than originally thought.
* Modified the scripted camera script so the focus camera script can
directly tell it where to move the camera.
* Modified the Focus camera script so it is easily configured to move
the camera in the way we want.
+ Created a new script called "SetCameraPosition". This is to give the
effect that the player has entered into a new area. The change is
instant.
* Added in Hao's new design for the swarmer pod. I also tweaked the
prefab and the script a bit to accomadate these changes.
* Added in Hao's new backgrounds.